@@ -8,13 +8,13 @@ namespace SongFeedReaders.DataflowAlternative
 {
     public class BlockResult<TOutput>
     {
-        public BlockResult(TOutput output, Exception exception = null)
+        public BlockResult(TOutput output, Exception? exception = null)
         {
             Output = output;
             Exception = exception;
         }
 
         public TOutput Output { get; private set; }
-        public Exception Exception { get; private set; }
+        public Exception? Exception { get; private set; }
     }
 }
