@@ -5,26 +5,47 @@ using System.Text;
 using System.Threading.Tasks;
 using WebUtilities;
 using SongFeedReaders.Logging;
+using SongFeedReaders.Data;
 
 namespace SongFeedReaders.Readers
 {
     public class PageReadResult
     {
-        private static FeedReaderLoggerBase _logger;
+        private static FeedReaderLoggerBase? _logger;
         public static FeedReaderLoggerBase Logger
         {
             get { return _logger ?? LoggingController.DefaultLogger; }
             set { _logger = value; }
         }
         public Uri Uri { get; private set; }
-        public List<ScrapedSong> Songs { get; private set; }
+        public List<ScrapedSong>? Songs { get; private set; }
+        /// <summary>
+        /// First song on the unfiltered page.
+        /// </summary>
+        public readonly ScrapedSong? FirstSong;
+        /// <summary>
+        /// Last song on the unfiltered page.
+        /// </summary>
+        public readonly ScrapedSong? LastSong;
+        /// <summary>
+        /// Total unfiltered songs.
+        /// </summary>
+        public readonly int SongsOnPage;
+
+        public bool IsLastPage { get; private set; }
+
         public PageErrorType PageError { get; private set; }
         public int Count { get { return Songs?.Count ?? 0; } }
-        public FeedReaderException Exception { get; private set; }
+        public FeedReaderException? Exception { get; private set; }
+
         private bool _successful;
         public bool Successful { get { return _successful && Exception == null; } }
-        public PageReadResult(Uri uri, List<ScrapedSong> songs)
+        public PageReadResult(Uri uri, List<ScrapedSong>? songs, ScrapedSong? firstSong, ScrapedSong? lastSong, int songsOnPage, bool isLastPage = false)
         {
+            IsLastPage = isLastPage;
+            FirstSong = firstSong;
+            LastSong = lastSong;
+            SongsOnPage = songsOnPage;
             Uri = uri;
             if (songs == null)
             {
@@ -36,8 +57,8 @@ namespace SongFeedReaders.Readers
             Songs = songs;
         }
 
-        public PageReadResult(Uri uri, List<ScrapedSong> songs, Exception exception, PageErrorType pageError)
-            : this(uri, songs)
+        public PageReadResult(Uri uri, List<ScrapedSong>? songs, ScrapedSong? firstSong, ScrapedSong? lastSong, int songsOnPage, Exception? exception, PageErrorType pageError, bool isLastPage = false)
+            : this(uri, songs, firstSong, lastSong, songsOnPage, isLastPage)
         {
 
             if (exception != null)
@@ -59,7 +80,12 @@ namespace SongFeedReaders.Readers
             }
         }
 
-        public static PageReadResult FromWebClientException(WebClientException ex, Uri requestUri)
+        public override string ToString()
+        {
+            return $"{Uri} | {Songs?.Count.ToString() ?? "<NULL>"}";
+        }
+
+        public static PageReadResult FromWebClientException(WebClientException? ex, Uri requestUri)
         {
             PageErrorType pageError = PageErrorType.SiteError;
             string errorText = string.Empty;
@@ -78,15 +104,26 @@ namespace SongFeedReaders.Readers
                         break;
                 }
             }
-            string message = $"{errorText} getting page {requestUri}.";
+            string message = $"{errorText} getting page '{requestUri}'.";
             Logger?.Debug(message);
             // No need for a stacktrace if it's one of these errors.
-            if (!(pageError == PageErrorType.Timeout || statusCode == 500))
+            if (!(pageError == PageErrorType.Timeout || statusCode == 500) && ex != null)
                 Logger?.Debug($"{ex.Message}\n{ex.StackTrace}");
-            return new PageReadResult(requestUri, null, new FeedReaderException(message, ex, FeedReaderFailureCode.PageFailed), pageError);
+            return new PageReadResult(requestUri, null, null, null, 0, new FeedReaderException(message, ex, FeedReaderFailureCode.PageFailed), pageError);
+        }
 
+        public static PageReadResult CancelledResult(Uri requestUri, OperationCanceledException ex)
+        {
+            return new PageReadResult(requestUri, new List<ScrapedSong>(), null, null, 0, ex, PageErrorType.Cancelled);
+        }
+
+        public static PageReadResult CancelledResult(Uri requestUri)
+        {
+            return new PageReadResult(requestUri, new List<ScrapedSong>(), null, null, 0, new OperationCanceledException(), PageErrorType.Cancelled);
         }
     }
+
+
     public static class PageErrorTypeExtensions
     {
         public static string ErrorToString(this PageErrorType pageError)
@@ -101,6 +138,8 @@ namespace SongFeedReaders.Readers
                     return "Site Error";
                 case PageErrorType.ParsingError:
                     return "Parsing Error";
+                case PageErrorType.PageOutOfRange:
+                    return "Page out of range";
                 case PageErrorType.Unknown:
                     return "Unknown Error";
                 default:
@@ -115,6 +154,7 @@ namespace SongFeedReaders.Readers
         SiteError = 2,
         ParsingError = 3,
         Cancelled = 4,
-        Unknown = 5
+        Unknown = 5,
+        PageOutOfRange = 6
     }
 }
