@@ -1,15 +1,12 @@
-﻿using System;
+﻿using SongFeedReaders.Logging;
+using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.IO;
-using System.Net;
-using SongFeedReaders.Logging;
-using WebUtilities;
-using System.Collections.Concurrent;
 using System.Threading;
-
+using System.Threading.Tasks;
+using WebUtilities;
+using SongFeedReaders.Services;
 namespace SongFeedReaders
 {
     /// <summary>
@@ -17,8 +14,54 @@ namespace SongFeedReaders
     /// </summary>
     public static class WebUtils
     {
-        private static FeedReaderLoggerBase _logger;
-        public static FeedReaderLoggerBase Logger
+        private static Uri beatSaverUri = null!;
+        private static Uri beatSaverApiUri = null!;
+        private static Uri beatSaverDownloadUri = null!;
+        private static Uri beatSaverDetailsFromKeyBaseUrl = null!;
+        private static Uri beatSaverDetailsFromHashBaseUrl = null!;
+
+        public static Uri BeatSaverUri
+        {
+            get => beatSaverUri;
+            set
+            {
+                if (value != beatSaverUri && value != null)
+                {
+                    beatSaverUri = value;
+                    beatSaverApiUri = new Uri($"{value.Scheme}://api.{value.Host}", UriKind.Absolute);
+                    beatSaverDownloadUri = new Uri($"{value.Scheme}://cdn.{value.Host}", UriKind.Absolute);
+                    beatSaverDetailsFromKeyBaseUrl = new Uri(beatSaverApiUri, "/maps/beatsaver/");
+                    beatSaverDetailsFromHashBaseUrl = new Uri(beatSaverApiUri, "/maps/hash/");
+                }
+            }
+        }
+
+        public static Uri BeatSaverApiUri => beatSaverApiUri;
+        public static Uri BeatSaverDownloadUri => beatSaverDownloadUri;
+        public static Uri BeatSaverDetailsFromKeyBaseUrl => beatSaverDetailsFromKeyBaseUrl;
+        public static Uri BeatSaverDetailsFromHashBaseUrl => beatSaverDetailsFromHashBaseUrl;
+        static WebUtils()
+        {
+            BeatSaverUri = new Uri("https://beatsaver.com");
+        }
+
+
+        private static SongInfoManager? _songInfoManager;
+        public static SongInfoManager SongInfoManager
+        {
+            get
+            {
+                if (_songInfoManager == null)
+                {
+                    _songInfoManager = new SongInfoManager();
+                    _songInfoManager.AddProvider<BeatSaverSongInfoProvider>("BeatSaverProvider", 500);
+                }
+                return _songInfoManager;
+            }
+        }
+
+        private static FeedReaderLoggerBase? _logger;
+        public static FeedReaderLoggerBase? Logger
         {
             get { return _logger ?? LoggingController.DefaultLogger; }
             set { _logger = value; }
@@ -45,19 +88,19 @@ namespace SongFeedReaders
             "/maps/plays",
             "/maps/hot",
             "/maps/rating",
-            "/maps/detail/*",
-            "/maps/by-hash/*",
+            "/maps/detail/",
+            "/maps/by-hash/",
             "/maps/uploader/*"
         };
 
-        public static string GetRateLimitedBase(string url)
+        public static string? GetRateLimitedBase(string url)
         {
             if (string.IsNullOrEmpty(url))
                 return null;
-            string baseUrl = null;
+            string? baseUrl = null;
             bool wildCard = false;
 
-            var route = RateLimitedRoutes.Where(r => url.Contains(r.EndsWith("*") ? r.Substring(0, r.Length - 1) : r)).SingleOrDefault();
+            string? route = RateLimitedRoutes.Where(r => url.Contains(r.EndsWith("*") ? r.Substring(0, r.Length - 1) : r)).SingleOrDefault();
             if (string.IsNullOrEmpty(route))
                 return string.Empty;
             if (wildCard = route.EndsWith("*"))
@@ -70,8 +113,8 @@ namespace SongFeedReaders
             if (wildCard)
             {
                 string addition = string.Empty;
-                var afterRoute = url.Substring(url.IndexOf(route) + route.Length);
-                var firstSlash = afterRoute.IndexOf("/");
+                string afterRoute = url.Substring(url.IndexOf(route) + route.Length);
+                int firstSlash = afterRoute.IndexOf("/");
                 if (firstSlash > 0)
                     addition = afterRoute.Substring(0, firstSlash);
                 else
@@ -81,11 +124,11 @@ namespace SongFeedReaders
             return baseUrl;
         }
 
-        public static string GetRateLimitedBase(Uri uri)
+        public static string? GetRateLimitedBase(Uri uri)
         {
             return GetRateLimitedBase(uri.ToString());
         }
-        private static IWebClient _webClient;
+        private static IWebClient? _webClient;
 
         /// <summary>
         /// Returns the WebClient, throws a <see cref="NullReferenceException"/> if WebClient is null.
@@ -105,12 +148,13 @@ namespace SongFeedReaders
             }
         }
 
+
         /// <summary>
         /// Stores known rate limit remaining time for BeatSaver's routes.
         /// Key is the base URL calculated by <see cref="GetRateLimitedBase(string)"/>.
         /// </summary>
-        public static ConcurrentDictionary<string, RateLimitPair> WaitForRateLimitDict = new ConcurrentDictionary<string, RateLimitPair>();
-        public class RateLimitPair
+        internal static ConcurrentDictionary<string, RateLimitPair> WaitForRateLimitDict = new ConcurrentDictionary<string, RateLimitPair>();
+        internal class RateLimitPair
         {
             public RateLimitPair(int callsRemaining, DateTime timeToReset)
             {
@@ -121,10 +165,10 @@ namespace SongFeedReaders
                 }
             }
 
-            public RateLimitPair(RateLimit rateLimit)
+            public RateLimitPair(RateLimit? rateLimit)
             {
-                var _resetTime = rateLimit?.TimeToReset ?? DateTime.Now;
-                var _callsRemaining = rateLimit?.CallsRemaining ?? 0;
+                DateTime _resetTime = rateLimit?.TimeToReset ?? DateTime.Now;
+                int _callsRemaining = rateLimit?.CallsRemaining ?? 0;
                 lock (_updateLock)
                 {
                     if (_resetTime >= TimeToReset)
@@ -135,12 +179,12 @@ namespace SongFeedReaders
                 }
             }
 
-            public void Update(RateLimit rateLimit)
+            public void Update(RateLimit? rateLimit)
             {
                 if (rateLimit == null)
                     return;
-                var _resetTime = rateLimit.TimeToReset;
-                var _callsRemaining = rateLimit.CallsRemaining;
+                DateTime _resetTime = rateLimit.TimeToReset;
+                int _callsRemaining = rateLimit.CallsRemaining;
                 lock (_updateLock)
                 {
                     if (_resetTime >= TimeToReset)
@@ -161,9 +205,9 @@ namespace SongFeedReaders
         /// <param name="cancellationToken"></param>
         /// <exception cref="TaskCanceledException"></exception>
         /// <returns></returns>
-        public static async Task WaitForRateLimit(string baseUrl, RateLimit rateLimit, CancellationToken cancellationToken)
+        public static async Task WaitForRateLimit(string? baseUrl, RateLimit? rateLimit, CancellationToken cancellationToken)
         {
-            var existing = UpdateRateLimit(baseUrl, rateLimit);
+            RateLimitPair? existing = UpdateRateLimit(baseUrl, rateLimit);
             if (existing == null || existing.CallsRemaining > 0)
                 return;
             TimeSpan delay = existing.TimeToReset - DateTime.Now;
@@ -178,11 +222,11 @@ namespace SongFeedReaders
             return;
         }
 
-        public static RateLimitPair UpdateRateLimit(string baseUrl, RateLimit rateLimit)
+        private static RateLimitPair? UpdateRateLimit(string? baseUrl, RateLimit? rateLimit)
         {
-            if (string.IsNullOrEmpty(baseUrl))
+            if (string.IsNullOrEmpty(baseUrl) || baseUrl == null)
                 return null;
-            var existing = WaitForRateLimitDict.GetOrAdd(baseUrl, (f) => new RateLimitPair(rateLimit));
+            RateLimitPair existing = WaitForRateLimitDict.GetOrAdd(baseUrl, (f) => new RateLimitPair(rateLimit));
             existing.Update(rateLimit);
             return existing;
         }
@@ -199,12 +243,12 @@ namespace SongFeedReaders
         public static async Task<IWebResponseMessage> GetBeatSaverAsync(Uri uri, CancellationToken cancellationToken, int maxSecondsToWait = 0, int retries = 5)
         {
             cancellationToken.ThrowIfCancellationRequested();
-            bool retry = false;
+            bool retry;
             int tries = 0;
-            IWebResponseMessage response = null;
-            string baseUrl = GetRateLimitedBase(uri)
-                ?? throw new ArgumentNullException(nameof(uri), "uri cannot be null for WebUtils.GetBeatSaverAsync");
-            await WaitForRateLimit(baseUrl, null, cancellationToken).ConfigureAwait(false); // Wait for an existing rate limit if it exists
+            IWebResponseMessage? response = null!;
+            //string baseUrl = GetRateLimitedBase(uri)
+            //    ?? throw new ArgumentNullException(nameof(uri), "uri cannot be null for WebUtils.GetBeatSaverAsync");
+            //await WaitForRateLimit(baseUrl, null, cancellationToken).ConfigureAwait(false); // Wait for an existing rate limit if it exists
             do
             {
                 cancellationToken.ThrowIfCancellationRequested();
@@ -212,54 +256,58 @@ namespace SongFeedReaders
                 try
                 {
                     response = await WebClient.GetAsync(uri, cancellationToken).ConfigureAwait(false);
-                    try
-                    {
-                        var rateLimit = ParseBeatSaverRateLimit(response.Headers);
-                        UpdateRateLimit(baseUrl, rateLimit);
-                    }
-                    catch (Exception ex)
-                    {
-                        Logger?.Debug($"Error parsing rate limit for {uri}: {ex.Message}");
-                    }
+                    //try
+                    //{
+                    //    RateLimit? rateLimit = ParseBeatSaverRateLimit(response.Headers);
+                    //    UpdateRateLimit(baseUrl, rateLimit);
+                    //}
+                    //catch (Exception ex)
+                    //{
+                    //    Logger?.Debug($"Error parsing rate limit for {uri}: {ex.Message}");
+                    //}
+                    response.EnsureSuccessStatusCode();
                 }
                 catch (WebClientException ex)
                 {
-                    var statusCode = ex.Response?.StatusCode ?? 0;
+                    int statusCode = ex.Response?.StatusCode ?? 0;
                     if (tries >= retries && (statusCode != 429 || statusCode != 0))
                         throw;
-                    var errorCode = ex.Response?.StatusCode ?? 0;
+                    int errorCode = ex.Response?.StatusCode ?? 0;
                     if (errorCode == 429 && tries < retries)
                     {
 
                         retry = true;
-                        var rateLimit = ParseBeatSaverRateLimit(ex.Response.Headers);
+                        tries++;
+                        Logger?.Warning($"BeatSaver responded with '(429) Too Many Requests' on URL '{uri.OriginalString}', retrying after 10s...");
+                        await Task.Delay(10000).ConfigureAwait(false);
+                        //RateLimit? rateLimit = ParseBeatSaverRateLimit(ex.Response?.Headers) ?? new RateLimit() { CallsPerReset = 50, CallsRemaining = 0, TimeToReset = DateTime.Now + TimeSpan.FromSeconds(20) };
 
-                        WaitForRateLimitDict.AddOrUpdate(baseUrl, u => new RateLimitPair(rateLimit), (url, rateLimitPair) =>
-                        {
-                            rateLimitPair.Update(rateLimit);
-                            return rateLimitPair;
-                        });
-                        if (rateLimit != null)
-                        {
-                            TimeSpan delay = new TimeSpan(0);
-                            var calcDelay = rateLimit.TimeToReset.Add(RateLimitPadding) - DateTime.Now;
-                            if (calcDelay > TimeSpan.Zero) // Make sure the delay is > 0
-                                delay = calcDelay;
-                            if (maxSecondsToWait != 0 && delay.TotalSeconds > maxSecondsToWait)
-                            {
-                                Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, delay of {(int)delay.TotalSeconds} seconds is too long, cancelling...");
-                                throw ex;
-                            }
-                            Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, retrying in {(int)delay.TotalSeconds} seconds");
-                            await Task.Delay(delay).ConfigureAwait(false);
-                            tries++;
-                            continue;
-                        }
-                        else
-                        {
-                            Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, could not parse rate limit, not retrying.");
-                            throw ex;
-                        }
+                        //WaitForRateLimitDict.AddOrUpdate(baseUrl, u => new RateLimitPair(rateLimit), (url, rateLimitPair) =>
+                        //{
+                        //    rateLimitPair.Update(rateLimit);
+                        //    return rateLimitPair;
+                        //});
+                        //if (rateLimit != null)
+                        //{
+                        //    TimeSpan delay = new TimeSpan(0);
+                        //    TimeSpan calcDelay = rateLimit.TimeToReset.Add(RateLimitPadding) - DateTime.Now;
+                        //    if (calcDelay > TimeSpan.Zero) // Make sure the delay is > 0
+                        //        delay = calcDelay;
+                        //    if (maxSecondsToWait != 0 && delay.TotalSeconds > maxSecondsToWait)
+                        //    {
+                        //        Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, delay of {(int)delay.TotalSeconds} seconds is too long, cancelling...");
+                        //        throw ex;
+                        //    }
+                        //    Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, retrying in {(int)delay.TotalSeconds} seconds");
+                        //    await Task.Delay(delay).ConfigureAwait(false);
+                        //    tries++;
+                        //    continue;
+                        //}
+                        //else
+                        //{
+                        //    Logger?.Warning($"Try {tries}: Rate limit exceeded on url, {uri.ToString()}, could not parse rate limit, not retrying.");
+                        //    throw ex;
+                        //}
                     }
                     else if (errorCode == 0 && tries < retries)
                     {
@@ -275,7 +323,7 @@ namespace SongFeedReaders
                         throw ex;
                     }
                 }
-                
+
             } while (retry);
             return response;
         }
@@ -307,7 +355,7 @@ namespace SongFeedReaders
         /// </summary>
         /// <param name="headers"></param>
         /// <returns></returns>
-        public static RateLimit ParseBeatSaverRateLimit(IDictionary<string, IEnumerable<string>> headers)
+        public static RateLimit? ParseBeatSaverRateLimit(IDictionary<string, IEnumerable<string>>? headers)
         {
             if (headers == null)
             {
@@ -351,8 +399,26 @@ namespace SongFeedReaders
             if (!IsInitialized || _webClient == null)
             {
                 _webClient = client;
+                _ = SongInfoManager;
                 IsInitialized = true;
             }
+        }
+
+
+
+        public static Uri GetDownloadUriByHash(string hash)
+        {
+            return new Uri(BeatSaverDownloadUri, hash.ToLower() + ".zip");
+        }
+
+        public static Uri GetBeatSaverDetailsByKey(string key)
+        {
+            return new Uri(BeatSaverDetailsFromKeyBaseUrl + key.ToLower());
+        }
+
+        public static Uri GetBeatSaverDetailsByHash(string hash)
+        {
+            return new Uri(BeatSaverDetailsFromHashBaseUrl + hash.ToLower());
         }
     }
 
@@ -360,6 +426,7 @@ namespace SongFeedReaders
     {
         public int CallsRemaining { get; set; }
         public DateTime TimeToReset { get; set; }
+        // public TimeSpan ResetAfter { get; set; }
         public int CallsPerReset { get; set; }
     }
 }
