@@ -21,23 +21,24 @@ namespace SongFeedReadersTests.BeatSaverReaderTests
         [TestMethod]
         public void Success()
         {
-            string pageText = File.ReadAllText(@"Data\BeatSaverListPage.json");
+            string path = Path.Combine("Data", "NewBeatSaver", "List", "Latest", "Latest_0.json");
+            string pageText = File.ReadAllText(path);
             Uri uri = null;
-            var songs = BeatSaverReader.ParseSongsFromPage(pageText, uri);
-            Assert.IsTrue(songs.Count == 10);
+            var songs = BeatSaverReader.ParseSongsFromPage(pageText, uri, true);
+            Assert.AreEqual(20, songs.Count);
             foreach (var song in songs)
             {
                 Assert.IsFalse(song.DownloadUri == null);
                 Assert.IsFalse(string.IsNullOrEmpty(song.Hash));
-                Assert.IsFalse(string.IsNullOrEmpty(song.MapperName));
+                Assert.IsFalse(string.IsNullOrEmpty(song.LevelAuthorName));
                 Assert.IsFalse(string.IsNullOrEmpty(song.RawData));
-                Assert.IsFalse(string.IsNullOrEmpty(song.SongName));
+                Assert.IsFalse(string.IsNullOrEmpty(song.Name));
             }
             var firstSong = JObject.Parse(songs.First().RawData);
-            string firstHash = firstSong["hash"]?.Value<string>();
-            Assert.IsTrue(firstHash == "27639680f92a9588b7cce843fc7aaa0f5dc720f8");
-            string firstUploader = firstSong["uploader"]?["username"]?.Value<string>();
-            Assert.IsTrue(firstUploader == "latte");
+            string firstHash = firstSong["versions"].First()["hash"]?.Value<string>();
+            Assert.AreEqual("e2512e6fbf85059d9fd9b429f62b2e618dd4d7e9", firstHash);
+            string firstUploader = firstSong["uploader"]?["name"]?.Value<string>();
+            Assert.AreEqual("itzrimuru", firstUploader);
         }
     }
 }
