@@ -6,6 +6,8 @@ using SongFeedReaders;
 using Newtonsoft.Json.Linq;
 using System;
 using SongFeedReaders.Readers.ScoreSaber;
+using System.Threading;
+using SongFeedReaders.Readers;
 
 namespace SongFeedReadersTests.ScoreSaberReaderTests
 {
@@ -22,10 +24,37 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
         {
             var reader = new ScoreSaberReader();
             int maxSongs = 100;
-            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeed.Trending) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.Trending) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
             var songList = reader.GetSongsFromFeed(settings);
             Assert.IsTrue(songList.Count == maxSongs);
             Assert.IsFalse(songList.Songs.Keys.Any(k => string.IsNullOrEmpty(k)));
+        }
+
+        [TestMethod]
+        public void CancelledImmediate()
+        {
+            var reader = new ScoreSaberReader();
+            var cts = new CancellationTokenSource();
+            cts.Cancel();
+            int maxSongs = 50;
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.LatestRanked) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
+            var result = reader.GetSongsFromFeed(settings, cts.Token);
+            Assert.IsFalse(result.Successful);
+            Assert.AreEqual(FeedResultError.Cancelled, result.ErrorCode);
+            cts.Dispose();
+        }
+
+        [TestMethod]
+        public void CancelledInProgress()
+        {
+            var reader = new ScoreSaberReader();
+            var cts = new CancellationTokenSource(500);
+            int maxSongs = 50;
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.LatestRanked) { MaxSongs = maxSongs, SongsPerPage = 10, RankedOnly = true };
+            var result = reader.GetSongsFromFeed(settings, cts.Token);
+            Assert.IsFalse(result.Successful);
+            Assert.AreEqual(FeedResultError.Cancelled, result.ErrorCode);
+            cts.Dispose();
         }
 
         [TestMethod]
@@ -33,7 +62,7 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
         {
             var reader = new ScoreSaberReader();
             int maxSongs = 50;
-            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeed.LatestRanked) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.LatestRanked) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
             var result = reader.GetSongsFromFeed(settings);
             Assert.IsTrue(result.Count == maxSongs);
             Assert.IsFalse(result.Songs.Keys.Any(k => string.IsNullOrEmpty(k)));
@@ -45,7 +74,7 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
         {
             var reader = new ScoreSaberReader();
             int maxSongs = 0;
-            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeed.TopRanked) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.TopRanked) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true };
             var songList = reader.GetSongsFromFeed(settings);
             Console.WriteLine($"{songList.Count} songs.");
             Assert.IsTrue(songList.Count >= 367);
@@ -57,7 +86,7 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
         {
             var reader = new ScoreSaberReader();
             int maxSongs = 30;
-            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeed.TopPlayed) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = false };
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.TopPlayed) { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = false };
             var songList = reader.GetSongsFromFeed(settings);
             Assert.IsTrue(songList.Count == maxSongs);
             Assert.IsFalse(songList.Songs.Keys.Any(k => string.IsNullOrEmpty(k)));
@@ -68,10 +97,10 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
         {
             var reader = new ScoreSaberReader();
             int maxSongs = 40;
-            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeed.Search)
+            var settings = new ScoreSaberFeedSettings((int)ScoreSaberFeedName.Search)
             { MaxSongs = maxSongs, SongsPerPage = 40, RankedOnly = true, SearchQuery = "Believer" };
             var songList = reader.GetSongsFromFeed(settings);
-            Assert.IsTrue(songList.Songs.Values.Any(s => s.SongName.ToLower().Contains("believer")));
+            Assert.IsTrue(songList.Songs.Values.Any(s => s.Name.ToLower().Contains("believer")));
             Assert.IsFalse(songList.Songs.Keys.Any(k => string.IsNullOrEmpty(k)));
         }
 
@@ -81,14 +110,14 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
             var reader = new ScoreSaberReader() { StoreRawData = true };
             var pageText = File.ReadAllText("Data\\ScoreSaberPage.json");
             Uri sourceUri = null;
-            var songList = reader.GetSongsFromPageText(pageText, sourceUri);
+            var songList = ScoreSaberFeed.GetSongsFromPageText(pageText, sourceUri, true);
             Assert.IsTrue(songList.Count == 50);
             var firstHash = "0597F8F7D8E396EBFEF511DC9EC98B69635CE532";
-            Assert.IsTrue(songList.Songs.First().Hash == firstHash);
-            var firstRawData = JToken.Parse(songList.Songs.First().RawData);
+            Assert.IsTrue(songList.First().Hash == firstHash);
+            var firstRawData = JToken.Parse(songList.First().RawData);
             Assert.IsTrue(firstRawData["uid"]?.Value<int>() == 143199);
             var lastHash = "F369747C6B54914DEAA163AAE85816BA5A8C1845";
-            Assert.IsTrue(songList.Songs.Last().Hash == lastHash);
+            Assert.IsTrue(songList.Last().Hash == lastHash);
         }
 
         [TestMethod]
@@ -97,14 +126,14 @@ namespace SongFeedReadersTests.ScoreSaberReaderTests
             var reader = new ScoreSaberReader() { StoreRawData = true };
             var pageText = File.ReadAllText("Data\\ScoreSaberPage.json");
             string url = Path.GetFullPath("Data\\ScoreSaberPage.json");
-            var songList = reader.GetSongsFromPageText(pageText, url);
+            var songList = ScoreSaberFeed.GetSongsFromPageText(pageText, new Uri(url), true);
             Assert.IsTrue(songList.Count == 50);
             var firstHash = "0597F8F7D8E396EBFEF511DC9EC98B69635CE532";
-            Assert.IsTrue(songList.Songs.First().Hash == firstHash);
-            var firstRawData = JToken.Parse(songList.Songs.First().RawData);
+            Assert.IsTrue(songList.First().Hash == firstHash);
+            var firstRawData = JToken.Parse(songList.First().RawData);
             Assert.IsTrue(firstRawData["uid"]?.Value<int>() == 143199);
             var lastHash = "F369747C6B54914DEAA163AAE85816BA5A8C1845";
-            Assert.IsTrue(songList.Songs.Last().Hash == lastHash);
+            Assert.IsTrue(songList.Last().Hash == lastHash);
         }
 
     }
