@@ -7,9 +7,12 @@ using System.Text;
 
 namespace WebUtilities.HttpClientWrapper
 {
+    /// <summary>
+    /// Wrapper for the response from <see cref="HttpClientWrapper"/>.
+    /// </summary>
     public class HttpResponseWrapper : IWebResponseMessage
     {
-        private HttpResponseMessage _response;
+        private HttpResponseMessage? _response;
         private int? _statusCodeOverride;
         /// <summary>
         /// Returns 0 if response was null and no status override was provided.
@@ -24,14 +27,19 @@ namespace WebUtilities.HttpClientWrapper
             }
         }
 
+        /// <inheritdoc/>
         public bool IsSuccessStatusCode { get { return _response?.IsSuccessStatusCode ?? false; } }
 
-        public string ReasonPhrase { get { return _response?.ReasonPhrase ?? Exception?.Message; } }
+        /// <inheritdoc/>
+        public string? ReasonPhrase { get { return _response?.ReasonPhrase ?? Exception?.Message; } }
 
+        /// <inheritdoc/>
         public Uri RequestUri { get; protected set; }
 
-        public Exception Exception { get; protected set; }
+        /// <inheritdoc/>
+        public Exception? Exception { get; protected set; }
 
+        /// <inheritdoc/>
         public IWebResponseMessage EnsureSuccessStatusCode()
         {
             try
@@ -42,7 +50,7 @@ namespace WebUtilities.HttpClientWrapper
                 {
                     HttpRequestException httpException;
                     if(((int)_response.StatusCode) > 0)
-                        httpException = new HttpRequestException($"The remove server returned an error: ({(int)_response.StatusCode}) {_response.ReasonPhrase}.");
+                        httpException = new HttpRequestException($"The remote server returned an error: ({(int)_response.StatusCode}) {_response.ReasonPhrase}.");
                     else
                         httpException = new HttpRequestException($"Error getting a response: {_response.ReasonPhrase}.");
                     var faultedResponse = new FaultedResponse(this);
@@ -54,22 +62,31 @@ namespace WebUtilities.HttpClientWrapper
             catch (HttpRequestException ex)
             {
                 var faultedResponse = new FaultedResponse(this);
-                _response.Dispose();
+                _response?.Dispose();
                 _response = null;
                 throw new WebClientException(ex.Message, ex, faultedResponse);
             }
             return this;
         }
 
-        public IWebResponseContent Content { get; protected set; }
+        /// <inheritdoc/>
+        public IWebResponseContent? Content { get; protected set; }
 
         private Dictionary<string, IEnumerable<string>> _headers;
+        /// <inheritdoc/>
         public ReadOnlyDictionary<string, IEnumerable<string>> Headers
         {
             get { return new ReadOnlyDictionary<string, IEnumerable<string>>(_headers); }
         }
 
-        public HttpResponseWrapper(HttpResponseMessage response, Uri requestUri, Exception exception = null, int? statusCodeOverride = null)
+        /// <summary>
+        /// Creates a new <see cref="HttpResponseWrapper"/> from the given <paramref name="response"/>.
+        /// </summary>
+        /// <param name="response"></param>
+        /// <param name="requestUri"></param>
+        /// <param name="exception"></param>
+        /// <param name="statusCodeOverride"></param>
+        public HttpResponseWrapper(HttpResponseMessage? response, Uri requestUri, Exception? exception = null, int? statusCodeOverride = null)
         {
             _response = response;
             _statusCodeOverride = statusCodeOverride;
@@ -91,6 +108,7 @@ namespace WebUtilities.HttpClientWrapper
         #region IDisposable Support
         private bool disposedValue = false; // To detect redundant calls
 
+        /// <inheritdoc/>
         protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
@@ -104,6 +122,7 @@ namespace WebUtilities.HttpClientWrapper
                     }
                     if (_response != null)
                     {
+                        // TODO: Should I be disposing of the response here? Not the content's responsibility?
                         _response.Dispose();
                         _response = null;
                     }
@@ -112,6 +131,7 @@ namespace WebUtilities.HttpClientWrapper
             }
         }
 
+        /// <inheritdoc/>
         public void Dispose()
         {
             Dispose(true);
