@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 
 namespace WebUtilities
 {
+    /// <summary>
+    /// An interface that defines a web client.
+    /// </summary>
     public interface IWebClient : IDisposable
     {
         /// <summary>
@@ -15,7 +18,7 @@ namespace WebUtilities
         /// <summary>
         /// The UserAgent string the client sends with request headers. Must start with "PRODUCT/VERSION" (No '\').
         /// </summary>
-        string UserAgent { get; }
+        string? UserAgent { get; }
         /// <summary>
         /// How the WebClient handles errors. TODO: May not be fully implemented.
         /// </summary>
@@ -25,7 +28,7 @@ namespace WebUtilities
         /// Should be in the format: Product/Version.
         /// </summary>
         /// <param name="userAgent"></param>
-        void SetUserAgent(string userAgent);
+        void SetUserAgent(string? userAgent);
 
         /// <summary>
         /// Send a GET request to the specified Uri as an asynchronous operation. If the server doesn't respond inside the provided timeout (milliseconds) or the provided CancellationToken is triggered, the operation is canceled.
@@ -87,7 +90,8 @@ namespace WebUtilities
         /// <returns></returns>
         Task<IWebResponseMessage> GetAsync(string url, CancellationToken cancellationToken);
         /// <summary>
-        /// Send a GET request to the specified Uri as an asynchronous operation. If the server doesn't respond inside the provided timeout (milliseconds) or the provided CancellationToken is triggered, the operation is canceled.
+        /// Send a GET request to the specified Uri as an asynchronous operation. 
+        /// If the server doesn't respond inside the provided timeout (milliseconds) or the provided CancellationToken is triggered, the operation is canceled.
         /// </summary>
         /// <param name="url"></param>
         /// <param name="timeout">Timeout in milliseconds</param>
@@ -102,10 +106,22 @@ namespace WebUtilities
 
     }
 
+    /// <summary>
+    /// The type of error handling the <see cref="IWebClient"/> uses.
+    /// </summary>
     public enum ErrorHandling
     {
+        /// <summary>
+        /// Any thrown exceptions are passed through to the caller.
+        /// </summary>
         ThrowOnException,
+        /// <summary>
+        /// Any web faults with throw a <see cref="WebClientException"/>.
+        /// </summary>
         ThrowOnWebFault,
+        /// <summary>
+        /// Any <see cref="Exception"/>s thrown are stored and returned in a <see cref="IWebResponse"/> with empty content.
+        /// </summary>
         ReturnEmptyContent
     }
 }
