@@ -9,10 +9,17 @@ using System.Threading;
 
 namespace WebUtilities.WebWrapper
 {
+    /// <summary>
+    /// Wrapper for the content of a <see cref="WebClientResponseWrapper"/>.
+    /// </summary>
     public class WebClientContent : IWebResponseContent
     {
-        private HttpWebResponse _response;
-        public WebClientContent(HttpWebResponse response)
+        private HttpWebResponse? _response;
+        /// <summary>
+        /// Creates a new <see cref="WebClientContent"/> from the <paramref name="response"/>.
+        /// </summary>
+        /// <param name="response"></param>
+        public WebClientContent(HttpWebResponse? response)
         {
             _response = response;
             ContentLength = _response?.ContentLength ?? 0;
@@ -28,12 +35,19 @@ namespace WebUtilities.WebWrapper
             }
         }
 
+        /// <summary>
+        /// Response headers.
+        /// </summary>
         protected Dictionary<string, IEnumerable<string>> _headers;
+
+        /// <summary>
+        /// A ReadOnlyDictionary of the response headers.
+        /// </summary>
         public ReadOnlyDictionary<string, IEnumerable<string>> Headers
         {
             get { return new ReadOnlyDictionary<string, IEnumerable<string>>(_headers); }
         }
-
+        /// <inheritdoc/>
         public string ContentType
         {
             get
@@ -46,33 +60,31 @@ namespace WebUtilities.WebWrapper
                 return cType;
             }
         }
-
+        /// <inheritdoc/>
         public long? ContentLength { get; protected set; }
-
+        /// <inheritdoc/>
         public async Task<byte[]> ReadAsByteArrayAsync()
         {
-            using (Stream stream = _response.GetResponseStream())
+            using (Stream stream = _response?.GetResponseStream() ?? throw new InvalidOperationException("There is no content to read."))
             {
                 using (MemoryStream memStream = new MemoryStream())
                 {
-                    await Task.Yield();
                     await stream.CopyToAsync(memStream, int.MaxValue).ConfigureAwait(false);
                     return memStream.ToArray();
                 }
             }
         }
-
+        /// <inheritdoc/>
         public Task<Stream> ReadAsStreamAsync()
         {
-            return Task.Run(() => _response?.GetResponseStream());
+            return Task.Run(() => _response?.GetResponseStream() ?? throw new InvalidOperationException("There is no content to read."));
         }
-
+        /// <inheritdoc/>
         public async Task<string> ReadAsStringAsync()
         {
-            using (Stream stream = _response?.GetResponseStream())
+            using (Stream stream = _response?.GetResponseStream() ?? throw new InvalidOperationException("There is no content to read."))
             using (var sr = new StreamReader(stream))
             {
-                await Task.Yield();
                 return await sr.ReadToEndAsync().ConfigureAwait(false);
             }
         }
@@ -82,6 +94,7 @@ namespace WebUtilities.WebWrapper
         /// </summary>
         /// <param name="filePath"></param>
         /// <param name="overwrite"></param>
+        /// <param name="cancellationToken"></param>
         /// <exception cref="ArgumentNullException">Thrown when content or the filename are null or empty.</exception>
         /// <exception cref="InvalidOperationException">Thrown when overwrite is false and a file at the provided path already exists.</exception>
         /// <exception cref="DirectoryNotFoundException">Thrown when the directory it's trying to save to doesn't exist.</exception>
@@ -91,8 +104,7 @@ namespace WebUtilities.WebWrapper
         /// <returns>Full path to the downloaded file</returns>
         public async Task<string> ReadAsFileAsync(string filePath, bool overwrite, CancellationToken cancellationToken)
         {
-            if (_response == null)
-                throw new ArgumentNullException(nameof(_response), "content cannot be null for HttpContent.ReadAsFileAsync");
+            HttpWebResponse response = _response ?? throw new InvalidOperationException("There is no content to read.");
             if (string.IsNullOrEmpty(filePath?.Trim()))
                 throw new ArgumentNullException(nameof(filePath), "filename cannot be null or empty for HttpContent.ReadAsFileAsync");
             string pathname = Path.GetFullPath(filePath);
@@ -101,19 +113,19 @@ namespace WebUtilities.WebWrapper
                 throw new InvalidOperationException(string.Format("File {0} already exists.", pathname));
             }
 
-            FileStream fileStream = null;
+            FileStream? fileStream = null;
             try
             {
                 fileStream = new FileStream(pathname, FileMode.Create, FileAccess.Write, FileShare.None);
-                var responseStream = _response.GetResponseStream();
+                var responseStream = response.GetResponseStream();
                 long expectedLength = 0;
-                if (_response.ContentLength > 0)
-                    expectedLength = _response.ContentLength;
+                if (response.ContentLength > 0)
+                    expectedLength = response.ContentLength;
                 // TODO: Timeouts don't seem to do anything.
                 //responseStream.ReadTimeout = 1;
                 //responseStream.WriteTimeout = 1;
                 string downloadedPath = string.Empty;
-                downloadedPath = await _response.GetResponseStream().CopyToAsync(fileStream, 81920, cancellationToken).ContinueWith(
+                downloadedPath = await response.GetResponseStream().CopyToAsync(fileStream, 81920, cancellationToken).ContinueWith(
                     (copyTask) =>
                     {
                         long fileStreamLength = fileStream.Length;
@@ -142,7 +154,7 @@ namespace WebUtilities.WebWrapper
 
         #region IDisposable Support
         private bool disposedValue = false; // To detect redundant calls
-
+        /// <inheritdoc/>
         protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
@@ -151,6 +163,7 @@ namespace WebUtilities.WebWrapper
                 {
                     if (_response != null)
                     {
+                        // TODO: Should I be disposing of the response here? Not the content's responsibility?
                         _response.Dispose();
                         _response = null;
                     }
@@ -158,7 +171,7 @@ namespace WebUtilities.WebWrapper
                 disposedValue = true;
             }
         }
-
+        /// <inheritdoc/>
         public void Dispose()
         {
             Dispose(true);
