@@ -6,15 +6,19 @@ using System.Text;
 
 namespace WebUtilities.WebWrapper
 {
+    /// <summary>
+    /// Response wrapper for <see cref="WebClientWrapper"/>.
+    /// </summary>
     public class WebClientResponseWrapper : IWebResponseMessage
     {
-        private HttpWebResponse _response;
-        private HttpWebRequest _request;
+        private HttpWebResponse? _response;
+        //private HttpWebRequest? _request;
         /// <summary>
         /// Returns 0 if response was null and no status override was provided.
         /// </summary>
         private int? _statusCodeOverride;
 
+        /// <inheritdoc/>
         public int StatusCode
         {
             get
@@ -25,15 +29,19 @@ namespace WebUtilities.WebWrapper
             }
         }
 
+        /// <inheritdoc/>
         public bool IsSuccessStatusCode
         {
             get { return (StatusCode >= 200) && (StatusCode <= 299); }
         }
 
-        public Exception Exception { get; protected set; }
+        /// <inheritdoc/>
+        public Exception? Exception { get; protected set; }
 
+        /// <inheritdoc/>
         public Uri RequestUri { get; protected set; }
 
+        /// <inheritdoc/>
         public IWebResponseMessage EnsureSuccessStatusCode()
         {
             if (_response == null)
@@ -46,7 +54,7 @@ namespace WebUtilities.WebWrapper
                 else if (Exception is WebException wException)
                     webException = wException;
                 else
-                    webException = new WebException($"The remove server returned an error: ({(int)_response.StatusCode}) {ReasonPhrase}.");
+                    webException = new WebException($"The remote server returned an error: ({(int)_response.StatusCode}) {ReasonPhrase}.");
                 var faultedResponse = new FaultedResponse(this);
                 _response.Dispose();
                 _response = null;
@@ -55,25 +63,34 @@ namespace WebUtilities.WebWrapper
             return this;
         }
 
-        public IWebResponseContent Content { get; protected set; }
+        /// <inheritdoc/>
+        public IWebResponseContent? Content { get; protected set; }
 
         private Dictionary<string, IEnumerable<string>> _headers;
+        /// <inheritdoc/>
         public ReadOnlyDictionary<string, IEnumerable<string>> Headers
         {
             get { return new ReadOnlyDictionary<string, IEnumerable<string>>(_headers); }
         }
 
+        /// <inheritdoc/>
         public string ReasonPhrase { get { return _response?.StatusDescription ?? Exception?.Message ?? "Unknown Error"; } }
-
-        public WebClientResponseWrapper(HttpWebResponse response, HttpWebRequest request, Exception exception = null, int? statusCodeOverride = null)
+        /// <summary>
+        /// Creates a new <see cref="WebClientResponseWrapper"/> with the given <see cref="HttpWebResponse"/> and <see cref="HttpWebRequest"/>.
+        /// </summary>
+        /// <param name="response"></param>
+        /// <param name="request"></param>
+        /// <param name="exception"></param>
+        /// <param name="statusCodeOverride"></param>
+        public WebClientResponseWrapper(HttpWebResponse? response, HttpWebRequest request, Exception? exception = null, int? statusCodeOverride = null)
         {
+            _ = request ?? throw new ArgumentNullException(nameof(request), $"Request cannot be null when creating a {nameof(WebClientResponseWrapper)}");
             _response = response;
-            _request = request;
             _statusCodeOverride = statusCodeOverride;
             Exception = exception;
-            RequestUri = request?.RequestUri;
+            RequestUri = request.RequestUri;
             if(response != null)
-                Content = new WebClientContent(_response);
+                Content = new WebClientContent(response);
             _headers = new Dictionary<string, IEnumerable<string>>();
             if (_response?.Headers != null)
             {
@@ -88,6 +105,10 @@ namespace WebUtilities.WebWrapper
         #region IDisposable Support
         private bool disposedValue = false; // To detect redundant calls
 
+        /// <summary>
+        /// Disposes of the <see cref="Content"/> and wrapped response, if they exist.
+        /// </summary>
+        /// <param name="disposing"></param>
         protected virtual void Dispose(bool disposing)
         {
             if (!disposedValue)
@@ -109,6 +130,7 @@ namespace WebUtilities.WebWrapper
             }
         }
 
+        /// <inheritdoc/>
         public void Dispose()
         {
             Dispose(true);
