@@ -2,37 +2,37 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.IO;
 using System.Threading;
-using System.Threading.Tasks;
 using WebUtilities;
 using WebUtilities.HttpClientWrapper;
+using WebUtilities.WebWrapper;
 
 namespace WebUtilitiesTests.HttpClientWrapperTests
 {
     [TestClass]
     public class WebClientContent_Tests
     {
-        private static readonly string TestOutputPath = Path.GetFullPath(@"Output\HttpClientContent");
+        private static readonly string TestOutputPath = Path.GetFullPath(@"Output\WebClientContent_Tests");
 
         [TestMethod]
         public void CanceledDownload()
         {
             IWebClient client = new HttpClientWrapper();
-            var uri = new Uri("http://releases.ubuntu.com/18.04.3/ubuntu-18.04.3-desktop-amd64.iso");
-            var cts = new CancellationTokenSource(500);
-            string directory = Path.Combine(TestOutputPath, "CanceledDownload");
-            Directory.CreateDirectory(directory);
-            string filePath = Path.Combine(directory, "CanceledDownload.iso");
-            
-            using(var response = client.GetAsync(uri, cts.Token).Result)
+            Uri uri = new Uri("http://releases.ubuntu.com/18.04.3/ubuntu-18.04.3-desktop-amd64.iso");
+            CancellationTokenSource cts = new CancellationTokenSource(700);
+            Directory.CreateDirectory(TestOutputPath);
+            string filePath = Path.Combine(TestOutputPath, "CanceledDownload.iso");
+
+            using (IWebResponseMessage response = client.GetAsync(uri).Result)
             {
                 if (!response.IsSuccessStatusCode)
                     Assert.Fail($"Error getting response from {uri}: {response.ReasonPhrase}");
                 try
                 {
-                    var thing = response.Content.ReadAsFileAsync(filePath, true, cts.Token).Result;
+                    CancellationTokenSource readCts = new CancellationTokenSource(100);
+                    string thing = response.Content.ReadAsFileAsync(filePath, true, readCts.Token).Result;
                     Assert.Fail("Didn't throw exception");
                 }
-                catch(AggregateException ex)
+                catch (AggregateException ex)
                 {
                     if (!(ex.InnerException is OperationCanceledException))
                         Assert.Fail("Wrong exception thrown.");
@@ -41,6 +41,37 @@ namespace WebUtilitiesTests.HttpClientWrapperTests
                 {
                     Assert.Fail("Wrong exception thrown.");
                 }
+            }
+            cts.Dispose();
+            client.Dispose();
+        }
+
+        [TestMethod]
+        public void CanceledBeforeResponse()
+        {
+            IWebClient client = new WebClientWrapper();
+            Uri uri = new Uri("http://releases.ubuntu.com/18.04.3/ubuntu-18.04.3-desktop-amd64.iso");
+            CancellationTokenSource cts = new CancellationTokenSource(100);
+            Directory.CreateDirectory(TestOutputPath);
+            string filePath = Path.Combine(TestOutputPath, "CanceledBeforeResponse.iso");
+            try
+            {
+                using (IWebResponseMessage response = client.GetAsync(uri, cts.Token).Result)
+                {
+                    if (!response.IsSuccessStatusCode)
+                        Assert.Fail($"Error getting response from {uri}: {response.ReasonPhrase}");
+                    Assert.Fail("Didn't throw exception");
+
+                }
+            }
+            catch (AggregateException ex)
+            {
+                if (!(ex.InnerException is OperationCanceledException))
+                    Assert.Fail("Wrong exception thrown.");
+            }
+            catch (Exception)
+            {
+                Assert.Fail("Wrong exception thrown.");
             }
             cts.Dispose();
             client.Dispose();
